+++ src/azure-cli/azure/cli/command_modules/resource/custom.py
@@ -284,9 +284,12 @@ def _deploy_arm_template_core(cli_ctx, resource_group_name,
                                       parameters=parameters, mode=mode, on_error_deployment=on_error_deployment)
 
     smc = get_mgmt_service_client(cli_ctx, ResourceType.MGMT_RESOURCE_RESOURCES, aux_subscriptions=aux_subscriptions)
+
     validation_result = smc.deployments.validate(resource_group_name=resource_group_name, deployment_name=deployment_name, properties=properties)
 
-    if validate_only or validation_result.error:
+    if validation_result and validation_result.error:
+        raise CLIError(validation_result.error)
+    if validate_only:
         return validation_result
 
     return sdk_no_wait(no_wait, smc.deployments.create_or_update, resource_group_name, deployment_name, properties)
@@ -367,7 +370,9 @@ def _deploy_arm_template_core_unmodified(cli_ctx, resource_group_name, template_
 
     validation_result = deployments_operation_group.validate(resource_group_name=resource_group_name, deployment_name=deployment_name, properties=properties)
 
-    if validate_only or validation_result.error:
+    if validation_result and validation_result.error:
+        raise CLIError(validation_result.error)
+    if validate_only:
         return validation_result
 
     return sdk_no_wait(no_wait, deployments_operation_group.create_or_update, resource_group_name, deployment_name, properties)
@@ -450,7 +455,9 @@ def _deploy_arm_template_at_subscription_scope(cli_ctx,
 
     validation_result = mgmt_client.validate_at_subscription_scope(deployment_name=deployment_name, properties=deployment_properties, location=deployment_location)
 
-    if validate_only or validation_result.error:
+    if validation_result and validation_result.error:
+        raise CLIError(validation_result.error)
+    if validate_only:
         return validation_result
 
     return sdk_no_wait(no_wait, mgmt_client.create_or_update_at_subscription_scope,
@@ -508,7 +515,9 @@ def _deploy_arm_template_at_resource_group(cli_ctx,
 
     validation_result = mgmt_client.validate(resource_group_name=resource_group_name, deployment_name=deployment_name, properties=deployment_properties)
 
-    if validate_only or validation_result.error:
+    if validation_result and validation_result.error:
+        raise CLIError(validation_result.error)
+    if validate_only:
         return validation_result
 
     return sdk_no_wait(no_wait, mgmt_client.create_or_update, resource_group_name,
@@ -561,7 +570,9 @@ def _deploy_arm_template_at_management_group(cli_ctx,
 
     validation_result = mgmt_client.validate_at_management_group_scope(group_id=management_group_id, deployment_name=deployment_name, properties=deployment_properties, location=deployment_location)
 
-    if validate_only or validation_result.error:
+    if validation_result and validation_result.error:
+        raise CLIError(validation_result.error)
+    if validate_only:
         return validation_result
 
     return sdk_no_wait(no_wait, mgmt_client.create_or_update_at_management_group_scope,
@@ -609,7 +620,9 @@ def _deploy_arm_template_at_tenant_scope(cli_ctx,
 
     validation_result = mgmt_client.validate_at_tenant_scope(deployment_name=deployment_name, properties=deployment_properties, location=deployment_location)
 
-    if validate_only or validation_result.error:
+    if validation_result and validation_result.error:
+        raise CLIError(validation_result.error)
+    if validate_only:
         return validation_result
 
     return sdk_no_wait(no_wait, mgmt_client.create_or_update_at_tenant_scope,
@@ -971,7 +984,6 @@ def export_group_as_template(
 
     result = rcf.resource_groups.export_template(resource_group_name, ['*'], options=options)
 
-    print(json.dumps(result.template, indent=2))
     # pylint: disable=no-member
     # On error, server still returns 200, with details in the error attribute
     if result.error:
@@ -982,6 +994,8 @@ def export_group_as_template(
             logger.warning(str(error))
         for detail in getattr(error, 'details', None) or []:
             logger.error(detail.message)
+
+    return result.template
 
 
 def create_application(cmd, resource_group_name,